back/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
asgiref==3.8.1
asttokens==3.0.0
asyncio==3.4.3
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
decorator==5.1.1
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
djangorestframework==3.15.2
drf-yasg==1.21.8
executing==2.2.0
frozenlist==1.5.0
idna==3.10
inflection==0.5.1
ipython==8.32.0
jedi==0.19.2
matplotlib-inline==0.1.7
multidict==6.1.0
packaging==24.2
parso==0.8.4
prompt_toolkit==3.0.50
propcache==0.2.1
pure_eval==0.2.3
Pygments==2.19.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
yarl==1.18.3
